Move compiler:build to gulp/tasks/compiler.ts
