+++ bokehjs/gulp/tasks/build.ts
@@ -2,7 +2,7 @@ import * as gulp from "gulp"
 import * as runSequence from "run-sequence"
 
 gulp.task("build", [], (cb: (arg?: any) => void) => {
-  runSequence(["scripts", "styles"], cb)
+  runSequence(["scripts", "styles", "compiler:build"], cb)
 })
 
 gulp.task("dev-build", [], (cb: (arg?: any) => void) => {
